torch-qa/qwen-requirements.txt
transformers_stream_generator==0.0.4